requirements.txt
# Framework principal
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0

# Base de datos y ORM
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
alembic==1.12.0

# Autenticación y seguridad
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
bcrypt==4.0.1

# Formularios
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Correo electrónico
Flask-Mail==0.9.1

# Generación de reportes
fpdf2==2.7.5
pandas==2.1.0
openpyxl==3.1.2
XlsxWriter==3.1.2

# Programación de tareas
schedule==1.2.0

# Utilidades
Pillow==10.0.0  # Para procesamiento de imágenes
python-dateutil==2.8.2
pytz==2023.3

# Desarrollo y pruebas
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0

# Producción
gunicorn==21.2.0
